added sorting

+++ school-managemnet/Controllers/StudentController.cs
@@ -1,27 +1,33 @@
 ﻿using SchoolManagement.BLL;
 using SchoolManagement.Models;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.EntityFrameworkCore;
 
 namespace SchoolManagement.Controllers
 {
     [Authorize]
     public class StudentController(IStudent db, ILogger<StudentController> logger) : Controller
     {
-        private readonly ILogger<StudentController> _logger = logger;
-
-        // [Authorize]
-        public async Task<IActionResult> Details(int id)
+        
+        public string NameSort { get; set; }
+        public string DateSort { get; set; }
+        public string CurrentFilter { get; set; }
+        public string CurrentSort { get; set; }
+        private string ErrorMessage { get; set; }
+        public async Task<IActionResult> Details(int? id)
         {
-            Student student = await db.GetById(id);
+            if(id is null)
+                return NotFound();
+            
+            var student = await db.GetById(id.Value);
             return View(student);
         }
 
         public async Task<IActionResult> DisplayEdited()
         {
             int id = HttpContext.Session.GetInt32("Id")?? 0;
-            if (id != null)
+            if (id != 0)
             {
                 Student stu = await db.GetById(id);
                 return View(stu);
@@ -44,7 +50,7 @@ namespace SchoolManagement.Controllers
 
         public async Task<IActionResult> CheckEmail(string email, int id)
         {
-            List<Student> result = await db.GetAll();
+            List<Student> result = await db.GetAll().ToListAsync();
 
             if (result.Count == 0 || (result.Count == 1 && result[0].Id == id))
             {
@@ -105,23 +111,67 @@ namespace SchoolManagement.Controllers
         }
 
         [Authorize]
-        public async Task<IActionResult> Index()
+        public async Task<IActionResult> Index(string sortOrder, string searchString)
         {
+            
             var s = User.Identity?.Name;
-            var model = await db.GetAll();
-            return View(model);
+
+            ViewData["NameSort"] = string.IsNullOrEmpty(sortOrder) ? "name_desc" : "";
+            ViewData["DateSort"] = sortOrder == "Date" ? "date_desc" : "Date";
+            var students =  db.GetAll();
+
+
+            students = sortOrder switch
+            {
+                "name_desc" => students.OrderByDescending(student => student.LastName),
+                "Date" => students.OrderBy(student => student.EnrollmentDate),
+                "date_desc" => students.OrderByDescending(student => student.EnrollmentDate),
+                _ => students.OrderBy(student => student.LastName)
+            };
+            
+            return View(await students.AsNoTracking().ToListAsync());
         }
 
         [HttpPost]
-        public async Task<IActionResult> Delete(Student student)
+        public async Task<IActionResult> Delete(int? id)
         {
-            await db.Delete(student.Id);
-            return RedirectToAction("Index");
+            if (id is null)
+                return NotFound();
+            
+            var student = await db.GetById(id.Value);
+
+            if (student is null)
+                return NotFound();
+
+            try
+            {
+                await db.Delete(student.Id);
+                return RedirectToAction(nameof(Index));
+            }
+            catch (DbUpdateException ex)
+            {
+                logger.LogError(ex, ErrorMessage);
+                return RedirectToAction(nameof(Delete), new { id, saveChangesError = true });
+            }
+
         }
 
-        public async Task<IActionResult> Delete(int id)
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
         {
-            return View(await db.GetById(id));
+            if(id is null)
+                return NotFound();
+
+            var student = await db.GetById(id.Value);
+
+            if (student is null)
+                return NotFound();
+
+            if (saveChangesError.GetValueOrDefault())
+            {
+                ErrorMessage = $"Delete {id} failed. Try again";
+                ViewData["ErrorMessage"] = ErrorMessage;
+            }
+            return View(student);
         }
         
     }
